fix(guitarpro): load GP5 files rejected by strict header parsing
- Beams are stored whenever any time-sig bit is set (flags & 0x03 != 0),
  not only when both are set; matches Python guitarpro and TuxGuitar.
  Fixes 2/4↔4/4 changes (Metallica - The Unforgiven).
- Triplet-feel byte read as u8; unknown values fall back to None like
  Guitar Pro itself. Fixes byte 0x80 (Iron Maiden - Fear Of The Dark).
- RSE instrument ints widened i16 -> i32 (their on-disk width); some
  exports use effect IDs > i16::MAX. Fixes 0x00FFFFFF effect number
  (Judas Priest - Breaking the Law).

+++ guitarpro/src/model/optimized/beat.rs
@@ -150,15 +150,16 @@ pub struct GpMixTableItem {
 pub struct GpMixTableChange {
     #[serde(default, skip_serializing_if = "Option::is_none")]
     pub instrument: Option<GpMixTableItem>,
-    // RSE instrument (GP5 only, often default)
+    // RSE instrument (GP5 only, often default). Stored as 4-byte ints in the
+    // legacy file, so keep 32-bit width — some exports use IDs larger than i16.
     #[serde(default)]
-    pub rse_instrument: i16,
+    pub rse_instrument: i32,
     #[serde(default)]
-    pub rse_unknown: i16,
+    pub rse_unknown: i32,
     #[serde(default)]
-    pub rse_sound_bank: i16,
+    pub rse_sound_bank: i32,
     #[serde(default)]
-    pub rse_effect_number: i16,
+    pub rse_effect_number: i32,
     #[serde(default, skip_serializing_if = "String::is_empty")]
     pub rse_effect_category: String,
     #[serde(default, skip_serializing_if = "String::is_empty")]
